refactored core folder functionalities to respective modules
