+++ backend/src/modules/core/container.ts
@@ -2,16 +2,16 @@ import {ContainerModule} from 'inversify';
 import {QuestionRepository} from '#root/shared/database/providers/mongo/repositories/QuestionRepository.js';
 import {ContextRepository} from '#root/shared/database/providers/mongo/repositories/ContextRepository.js';
 import {AnswerRepository} from '#root/shared/database/providers/mongo/repositories/AnswerRepository.js';
-import {AiService} from './services/AiService.js';
+import {AiService} from '../ai/services/AiService.js';
 import {QuestionSubmissionRepository} from '#root/shared/database/providers/mongo/repositories/SubmissionRepository.js';
 import {CommentRepository} from '#root/shared/database/providers/mongo/repositories/CommentRespository.js';
 import {CommentController} from '../../modules/comment/controllers/CommentController.js';
 import {CommentService} from '../../modules/comment/services/CommentService.js';
-import { UserController } from './controllers/UserController.js';
-import { UserService } from './services/UserService.js';
+import { UserController } from '../user/controllers/UserController.js';
+import { UserService } from '../user/services/UserService.js';
 import { RequestRepository } from '#root/shared/database/providers/mongo/repositories/RequestRepository.js';
-import { NotificationController } from './controllers/NotificationController.js';
-import { NotificationService } from './services/NotificationService.js';
+import { NotificationController } from '../notification/controllers/NotificationController.js';
+import { NotificationService } from '../notification/services/NotificationService.js';
 import { NotificationRepository } from '#root/shared/database/providers/mongo/repositories/NotificationRepository.js';
 import { ReviewRepository } from '#root/shared/database/providers/mongo/repositories/ReviewRepository.js';
 import { AnswerService } from '../answer/services/AnswerService.js';
