+++ backend/src/shared/database/interfaces/INotificationRepository.ts
@@ -1,5 +1,5 @@
 
-import { NotificationResponse } from '#root/modules/core/classes/validators/NotificationValidators.js';
+import { NotificationResponse } from '#root/modules/notification/validators/NotificationValidators.js';
 import { INotification, ISubscription } from '#root/shared/interfaces/models.js';
 import {ClientSession, ObjectId} from 'mongodb';
